Update index.js
I don't see the point of storing the lastpubmsg as it inhibits the user request to re-issue the command. This causes issues when you consider that external devices such LED controllers can change the colour without using mqtthing. Example: Siri changes to red, LED remote changes to blue, Siri can't change the LED back to red.

mqttthing code could check the status of the led RGB and determine it has really changed but the easier approach is remove the condition on line 792

Same changes can apply to HSVLight on line 414

Thanks for your consideration and your great work!